Align names of gRPC methods
